pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "cafeapi"
version = "0.1.0"
[tool.poetry.dependencies]
"discord.py" = "^1.3"
python = "^3.7"
jishaku = "^1.18"
python-dotenv = "^0.13.0"
flask = "^1.1"
[tool.poetry.dev-dependencies]
